pyproject.toml
[project]
name = "capstone-project-agentic-ai-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Rajat Khandelwal",email = "rajatkhandelwal191@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langgraph (>=1.0.9,<2.0.0)",
    "streamlit (>=1.54.0,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "pymupdf (>=1.27.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "pypdf (>=6.7.3,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"